requirements.txt
click==8.0.3
colorama==0.4.4
DateTime==4.3
Flask==2.0.2
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pytz==2021.3
SQLAlchemy==1.4.29
Werkzeug==2.0.2
zope.interface==5.4.0
python-dotenv==0.19.2
